\name{LDOD-package}
\alias{LDOD-package}
\alias{LDOD}
\docType{package}
\title{
Finding Locally D-optimal optimal designs for some nonlinear
and generalized linear models.
}
\description{
This package provides functions for Finding Locally D-optimal
designs for Logistic, Negative Binomial, Poisson, Michaelis-Menten, Exponential,
Log-Linear, Emax, Richards, Weibull and Inverse Quadratic regression models
and also functions for auto-constructing Fisher information matrix and
Frechet derivative based on some input variables and without user-interfere. 
}
\details{
\tabular{ll}{
Package: \tab LDOD\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-02-24\cr
License: \tab GPL (>=2)\cr
}
}
\author{
Ehsan Masoudi, Majid Sarmad and Hooshang Talebi

Maintainer: Ehsan Masoudi <esn_mud@yahoo.com>

}
\references{
Masoudi, E., Sarmad, M. and Talebi, H. 2012, An Almost General Code in R to Find Optimal Design, In Proceedings of the 1st ISM International Statistical Conference 2012, 292-297.

}

